refactor: quick registration section separated into different components.
