+++ src/app/admin/events/check-in/[eventDayId]/page.tsx
@@ -1,17 +1,17 @@
 import type { Metadata } from "next";
 import { cookies } from "next/headers";
 import { Suspense } from "react";
+import QuickOnsiteRegistrationCard from "@/app/admin/events/check-in/[eventDayId]/_components/quick-registration/QuickOnsiteRegistrationCard";
 import tryCatch from "@/lib/server/tryCatch";
 import { parseStringParam } from "@/lib/utils";
 import { getEventDayDetails } from "@/server/events/queries/getEventDayDetails";
 import { getAllMembers } from "@/server/members/queries/getAllMembers";
 import { getEventRegistrationList } from "@/server/registration/queries/getEventRegistrationList";
 import CheckInDataDialog from "./_components/CheckInDataDialog";
+import CheckInPageLoading from "./_components/CheckInPageLoading";
 import EventDayDetails from "./_components/EventDayDetails";
-import QuickOnsiteRegistrationCard from "./_components/QuickOnsiteRegistrationCard";
 import QRCodeScanner from "./_components/qr-scanning/QRCodeScanner";
 import CheckInRegistrationPanel from "./_components/registration-list/CheckInRegistrationPanel";
-import CheckInPageLoading from "./loading";
 
 export const metadata: Metadata = {
   title: "Event Check-In | Admin",
